Added a license file using the BSD-3-Clause (new) license template
Added a license file using the BSD-3-Clause (new) license template. Retained the author's names and changed copyright date to 2007-2020.

Currently, the project's license terms are available in the COPYING.md file, which is not recognized as the BSD-3-Clause license template by GitHub. The purpose of this commit is to make it more obvious for developers, who browse the repository, to recognize the license.